rebase: sync with the upstream

+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesSlave.java
@@ -1,11 +1,11 @@
 package org.csanchez.jenkins.plugins.kubernetes;
 
 import java.io.IOException;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import org.apache.commons.lang.RandomStringUtils;
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.durabletask.executors.OnceRetentionStrategy;
 import org.jvnet.localizer.Localizable;
@@ -19,14 +19,15 @@ import hudson.model.Label;
 import hudson.model.Node;
 import hudson.model.TaskListener;
 import hudson.slaves.AbstractCloudSlave;
+import hudson.slaves.Cloud;
 import hudson.slaves.JNLPLauncher;
-import hudson.slaves.NodeProperty;
 import hudson.slaves.OfflineCause;
 import hudson.slaves.RetentionStrategy;
 import io.fabric8.kubernetes.api.model.DoneablePod;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.client.dsl.ClientPodResource;
+import io.fabric8.kubernetes.client.dsl.PodResource;
+import jenkins.model.Jenkins;
 
 /**
  * @author Carlos Sanchez carlos@apache.org
@@ -36,35 +37,33 @@ public class KubernetesSlave extends AbstractCloudSlave {
     private static final Logger LOGGER = Logger.getLogger(KubernetesSlave.class.getName());
 
     private static final long serialVersionUID = -8642936855413034232L;
+    private static final String DEFAULT_AGENT_PREFIX = "jenkins-agent";
 
     /**
      * The resource bundle reference
      */
     private final static ResourceBundleHolder HOLDER = ResourceBundleHolder.get(Messages.class);
 
-    // private final Pod pod;
+    private final String cloudName;
 
-    private transient final KubernetesCloud cloud;
-
-    @DataBoundConstructor
     public KubernetesSlave(PodTemplate template, String nodeDescription, KubernetesCloud cloud, String labelStr)
             throws Descriptor.FormException, IOException {
-        super(getSlaveName(template),
-                nodeDescription,
-                template.getRemoteFs(),
-                1,
-                Node.Mode.NORMAL,
-                labelStr == null ? null : labelStr,
-                new JNLPLauncher(),
-                new OnceRetentionStrategy(cloud.getRetentionTimeout()),
-                template.getNodeProperties());
-        if (template.getAlways()) {
-            this.setRetentionStrategy(RetentionStrategy.INSTANCE);
-            this.setMode(Node.Mode.EXCLUSIVE);
-            LOGGER.warning("use always retention strategy, pod will be alive always");
-        }
-        // this.pod = pod;
-        this.cloud = cloud;
+
+
+        this(template, nodeDescription, cloud, labelStr, new OnceRetentionStrategy(cloud.getRetentionTimeout()));
+    }
+
+    public String getDescription() {
+        HashMap<String, String> des = new HashMap<>();
+        des.put("name", this.name);
+        des.put("nodeDescription", this.getNodeDescription());
+        des.put("remoteFs", this.getRemoteFS());
+        des.put("executors", String.valueOf(this.getNumExecutors()));
+        des.put("mode", this.getMode().getName());
+        des.put("label", this.getLabelString());
+        des.put("retentionStrategy", this.getRetentionStrategy().getClass().getName());
+        return des.toString();
+
     }
 
     public String getDescription() {
@@ -82,14 +81,29 @@ public class KubernetesSlave extends AbstractCloudSlave {
     @Deprecated
     public KubernetesSlave(PodTemplate template, String nodeDescription, KubernetesCloud cloud, Label label)
             throws Descriptor.FormException, IOException {
+        this(template, nodeDescription, cloud, label.toString(), new OnceRetentionStrategy(cloud.getRetentionTimeout())) ;
+    }
+
+    @Deprecated
+    public KubernetesSlave(PodTemplate template, String nodeDescription, KubernetesCloud cloud, String labelStr,
+                           RetentionStrategy rs)
+            throws Descriptor.FormException, IOException {
+        this(template, nodeDescription, cloud.name, labelStr, rs);
+    }
+
+    @DataBoundConstructor
+    public KubernetesSlave(PodTemplate template, String nodeDescription, String cloudName, String labelStr,
+                           RetentionStrategy rs)
+            throws Descriptor.FormException, IOException {
+
         super(getSlaveName(template),
                 nodeDescription,
                 template.getRemoteFs(),
                 1,
                 Node.Mode.NORMAL,
-                label == null ? null : label.toString(),
+                labelStr == null ? null : labelStr,
                 new JNLPLauncher(),
-                new OnceRetentionStrategy(cloud.getRetentionTimeout()),
+                rs,
                 template.getNodeProperties());
         if (template.getAlways()) {
             this.setRetentionStrategy(RetentionStrategy.INSTANCE);
@@ -98,20 +112,28 @@ public class KubernetesSlave extends AbstractCloudSlave {
         }
 
         // this.pod = pod;
-        this.cloud = cloud;
+        this.cloudName = cloudName;
+    }
+
+    public String getCloudName() {
+        return cloudName;
+    }
+
+    public Cloud getCloud() {
+        return Jenkins.getInstance().getCloud(getCloudName());
     }
 
     static String getSlaveName(PodTemplate template) {
-        String hex = Long.toHexString(System.nanoTime());
+        String randString = RandomStringUtils.random(5, "bcdfghjklmnpqrstvwxz0123456789");
         String name = template.getName();
         if (StringUtils.isEmpty(name)) {
-            return hex;
+            return String.format("%s-%s", DEFAULT_AGENT_PREFIX,  randString);
         }
         // no spaces
-        name = template.getName().replace(" ", "-").toLowerCase();
-        // keep it under 256 chars
-        name = name.substring(0, Math.min(name.length(), 256 - hex.length()));
-        return String.format("%s-%s", name, hex);
+        name = name.replaceAll("[ _]", "-").toLowerCase();
+        // keep it under 63 chars (62 is used to account for the '-')
+        name = name.substring(0, Math.min(name.length(), 62 - randString.length()));
+        return String.format("%s-%s", name, randString);
     }
 
     @Override
@@ -125,19 +147,44 @@ public class KubernetesSlave extends AbstractCloudSlave {
 
         Computer computer = toComputer();
         if (computer == null) {
-            LOGGER.log(Level.SEVERE, "Computer for slave is null: {0}", name);
+            String msg = String.format("Computer for slave is null: %s", name);
+            LOGGER.log(Level.SEVERE, msg);
+            listener.fatalError(msg);
+            return;
+        }
+
+        if (getCloudName() == null) {
+            String msg = String.format("Cloud name is not set for slave, can't terminate: %s", name);
+            LOGGER.log(Level.SEVERE, msg);
+            listener.fatalError(msg);
             return;
         }
 
         try {
-            KubernetesClient client = cloud.connect();
-            ClientPodResource<Pod, DoneablePod> pods = client.pods().withName(name);
+            Cloud cloud = getCloud();
+            if (cloud == null) {
+                String msg = String.format("Slave cloud no longer exists: %s", getCloudName());
+                LOGGER.log(Level.WARNING, msg);
+                listener.fatalError(msg);
+                return;
+            }
+            if (!(cloud instanceof KubernetesCloud)) {
+                String msg = String.format("Slave cloud is not a KubernetesCloud, something is very wrong: %s",
+                        getCloudName());
+                LOGGER.log(Level.SEVERE, msg);
+                listener.fatalError(msg);
+                return;
+            }
+            KubernetesClient client = ((KubernetesCloud) cloud).connect();
+            PodResource<Pod, DoneablePod> pods = client.pods().withName(name);
             pods.delete();
-            LOGGER.log(Level.INFO, "Terminated Kubernetes instance for slave {0}", name);
+            String msg = String.format("Terminated Kubernetes instance for slave %s", name);
+            LOGGER.log(Level.INFO, msg);
+            listener.getLogger().println(msg);
             computer.disconnect(OfflineCause.create(new Localizable(HOLDER, "offline")));
             LOGGER.log(Level.INFO, "Disconnected computer {0}", name);
         } catch (Exception e) {
-            LOGGER.log(Level.SEVERE, "Failure to terminate instance for slave " + name, e);
+            LOGGER.log(Level.SEVERE, "Failed to terminate pod for slave " + name, e);
         }
     }
